delete field modify test
